Add files via upload
Predictive Modelling for Attorney Involvement.  
The objective of this project is to develop a model that predicts whether an attorney will be involved in a claim based on various claim-related factors.
We successfully built a machine learning model to predict attorney involvement in insurance claims. The Logistic Regression model was selected for its accuracy, stability, and interpretability.